Rename: 패키지 구조 개선을 위한 패키지 리네이밍
