+++ queosk/src/main/java/com/bttf/queosk/entity/Restaurant.java
@@ -1,8 +1,9 @@
 package com.bttf.queosk.entity;
 
 import com.bttf.queosk.config.baseEntity.BaseTimeEntity;
-import com.bttf.queosk.dto.OperationStatus;
-import com.bttf.queosk.dto.RestaurantCategory;
+
+import com.bttf.queosk.dto.enumerate.OperationStatus;
+import com.bttf.queosk.dto.enumerate.RestaurantCategory;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
