--- queosk/src/main/resources/application.yml
@@ -1,23 +1,0 @@
-spring:
-  h2:
-    console:
-      enabled: true
-      path: /h2-console
-
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:test
-    username: sa
-    password:
-
-  redis:
-    host: localhost
-    port: 6379
-
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    properties:
-      hibernate:
-        globally_quoted_identifiers: true
-    show-sql: true
